requirements.txt
google-genai
python-dotenv
typer[all]
rich
pyyaml
